+++ src/main/java/com/example/umc8th/controller/ArticleController.java
@@ -3,19 +3,14 @@ package com.example.umc8th.controller;
 import com.example.umc8th.converter.ReplyConverter;
 import com.example.umc8th.dto.ArticleRequestDTO;
 import com.example.umc8th.dto.ArticleResponseDTO;
-import com.example.umc8th.dto.ReplyRequestDTO;
 import com.example.umc8th.dto.ReplyResponseDTO;
 import com.example.umc8th.entity.Article;
-import com.example.umc8th.entity.Reply;
 import com.example.umc8th.global.apiPayload.CustomResponse;
-import com.example.umc8th.service.ArticleCommandService;
-import com.example.umc8th.service.ArticleDeleteService;
-import com.example.umc8th.service.ArticleQueryService;
-import com.example.umc8th.service.ArticleUpdateService;
+import com.example.umc8th.service.command.ArticleCommandService;
+import com.example.umc8th.service.query.ArticleQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -23,8 +18,6 @@ import java.util.List;
 public class ArticleController {
     private final ArticleQueryService articleQueryService;
     private final ArticleCommandService articleCommandService;
-    private final ArticleUpdateService articleUpdateService;
-    private final ArticleDeleteService articleDeleteService;
 
     @PostMapping("/articles")
     public CustomResponse<ArticleResponseDTO.CreateArticleResponseDTO> createArticle(@RequestBody ArticleRequestDTO.CreateArticleDTO dto){
@@ -47,20 +40,20 @@ public class ArticleController {
 
     @PutMapping("/articles/{articleId}")
     public CustomResponse<ArticleResponseDTO.ArticlePreviewDTO> putArticles(@PathVariable("articleId") Long articleId, @RequestBody ArticleRequestDTO.UpdateArticleDTO dto){
-        Article article = articleUpdateService.updateArticle(articleId, dto);
+        Article article = articleCommandService.updateArticle(articleId, dto);
         return CustomResponse.ok(ArticleResponseDTO.ArticlePreviewDTO.from(article));
     }
 
     @PatchMapping("/articles/{articleId}")
     public CustomResponse<ArticleResponseDTO.ArticlePreviewDTO> patchArticles(@PathVariable("articleId") Long articleId, @RequestBody ArticleRequestDTO.UpdateArticleContentDTO dto){
-        Article article = articleUpdateService.updateArticleContent(articleId, dto);
+        Article article = articleCommandService.updateArticleContent(articleId, dto);
         return CustomResponse.ok(ArticleResponseDTO.ArticlePreviewDTO.from(article));
     }
 
 
     @DeleteMapping("/articles/{articleId}")
     public CustomResponse<ReplyResponseDTO.DeleteReplyDTO> deleteArticles(@PathVariable("articleId") Long articleId){
-        Long deleted = articleDeleteService.deleteArticle(articleId);
+        Long deleted = articleCommandService.deleteArticle(articleId);
         return CustomResponse.ok(ReplyConverter.toDeleteReplyDTO(deleted));
     }
 }
