+++ src/main/java/com/example/umc8th/controller/ReplyController.java
@@ -5,10 +5,8 @@ import com.example.umc8th.dto.ReplyRequestDTO;
 import com.example.umc8th.dto.ReplyResponseDTO;
 import com.example.umc8th.entity.Reply;
 import com.example.umc8th.global.apiPayload.CustomResponse;
-import com.example.umc8th.service.ReplyCommandService;
-import com.example.umc8th.service.ReplyQueryService;
-import com.example.umc8th.service.impl.ReplyDeleteServiceImpl;
-import com.example.umc8th.service.impl.ReplyUpdateServiceImpl;
+import com.example.umc8th.service.command.ReplyCommandService;
+import com.example.umc8th.service.query.ReplyQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,8 +17,6 @@ import java.util.List;
 public class ReplyController {
     private final ReplyQueryService replyQueryService;
     private final ReplyCommandService replyCommandService;
-    private final ReplyUpdateServiceImpl replyUpdateServiceImpl;
-    private final ReplyDeleteServiceImpl replyDeleteServiceImpl;
 
     @PostMapping("/replies")
     public CustomResponse<ReplyResponseDTO.CreateReplyResponseDTO> createReply(@RequestBody ReplyRequestDTO.CreateReplyDTO dto){
@@ -42,13 +38,13 @@ public class ReplyController {
 
     @PatchMapping("/replies/{replyId}")
     public CustomResponse<ReplyResponseDTO.ReplyPreviewDTO> patchReplies(@PathVariable("replyId") Long replyId, @RequestBody ReplyRequestDTO.UpdateReplyDTO dto){
-        Reply reply = replyUpdateServiceImpl.updateReply(replyId, dto);
+        Reply reply = replyCommandService.updateReply(replyId, dto);
         return CustomResponse.ok(ReplyConverter.toReplyPreviewDTO(reply));
     }
 
     @DeleteMapping("/replies/{replyId}")
     public CustomResponse<ReplyResponseDTO.DeleteReplyDTO> deleteReplies(@PathVariable("replyId") Long replyId){
-        Long replied = replyDeleteServiceImpl.deleteReply(replyId);
+        Long replied = replyCommandService.deleteReply(replyId);
         return CustomResponse.ok(ReplyConverter.toDeleteReplyDTO(replied));
     }
