+++ src/main/java/com/example/umc8th/entity/Article.java
@@ -15,7 +15,6 @@ import java.util.List;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@EntityListeners(AuditingEntityListener.class)
 @Getter
 public class Article extends BaseEntity {
     @Id
