--- src/main/java/com/example/umc8th/service/ArticleCommandService.java
@@ -1,8 +1,0 @@
-package com.example.umc8th.service;
-
-import com.example.umc8th.dto.ArticleRequestDTO;
-import com.example.umc8th.entity.Article;
-
-public interface ArticleCommandService {
-    Article createArticle(ArticleRequestDTO.CreateArticleDTO dto);
-}
