--- src/main/java/com/example/umc8th/service/ArticleDeleteService.java
@@ -1,5 +1,0 @@
-package com.example.umc8th.service;
-
-public interface ArticleDeleteService {
-    public Long deleteArticle(Long id);
-}
