--- src/main/java/com/example/umc8th/service/ReplyDeleteService.java
@@ -1,5 +1,0 @@
-package com.example.umc8th.service;
-
-public interface ReplyDeleteService {
-    public Long deleteReply(Long id);
-}
