+++ src/main/java/com/example/umc8th/service/query/ArticleQueryService.java
@@ -1,4 +1,4 @@
-package com.example.umc8th.service;
+package com.example.umc8th.service.query;
 
 import com.example.umc8th.entity.Article;
