+++ src/main/java/com/example/umc8th/service/query/ReplyQueryService.java
@@ -1,7 +1,6 @@
-package com.example.umc8th.service;
+package com.example.umc8th.service.query;
 
 
-import com.example.umc8th.entity.Article;
 import com.example.umc8th.entity.Reply;
 
 import java.util.List;
