+++ src/main/java/com/example/umc8th/service/query/impl/ReplyQueryServiceImpl.java
@@ -1,8 +1,8 @@
-package com.example.umc8th.service.impl;
+package com.example.umc8th.service.query.impl;
 
 import com.example.umc8th.entity.Reply;
 import com.example.umc8th.repository.ReplyRepository;
-import com.example.umc8th.service.ReplyQueryService;
+import com.example.umc8th.service.query.ReplyQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
